add include directive to the Fluentd configuration
This change adds the @include directive to the Fluentd configuration and creates a possibility to add some custom configuration blocks to the already existing configuration.
This would be extremly useful for the cases where one wants to manipulate or filter the events or maybe even wants to forward them to some external logging system apart from `log-cache-syslog`.

Co-authored-by: Jakob Schmid <jakob.schmid@sap.com>